refactor: move KeyActionMapper to src/utils
